backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
fsrs==3.1.0
pytest==8.4.2
pytest-asyncio==0.25.3
testcontainers[postgres]==4.13.3
httpx==0.28.1
openai==1.66.3
python-multipart==0.0.20
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
pydantic-settings==2.7.1
python-dotenv==1.2.1
